# omd_api.openapi.yaml
openapi: 3.1.0
info:
  title: 餐饮店菜单管理 API | v1
  version: 1.0.0
  description: |
    面向**餐饮店老板端 APP** 的菜单管理接口契约。  
    **本文件仅使用 `GET`（查询）与 `POST`（提交/变更）两种 HTTP 方法。**

    > **文档阅读说明**：各接口下方直接写出路径/Query/Body 参数与返回字段；**公共请求头见第二节**，单个接口不再重复列出。

    ---

    ## 一、环境域名

    | 环境 | Base URL | 说明 |
    |------|----------|------|
    | **生产（prod）** | `https://api.offergo.com/restaurant/v1` | 线上正式环境 |
    | **开发（dev）** | `https://dev-api.offergo.com/restaurant/v1` | 联调、测试环境 |

    ---

    ## 二、统一请求头（全局，各接口默认遵循）

    除标注「无需鉴权」的接口外，均需携带 **Authorization**。所有 **POST** Body 为 JSON。

    | Header 名称 | 必填 | 示例 | 作用 |
    |-------------|------|------|------|
    | `Content-Type` | POST 必填 | `application/json` | 请求体类型 |
    | `Accept` | 建议 | `application/json` | 期望响应为 JSON |
    | `Authorization` | 除登录外必填 | `Bearer eyJhbGciOi...` | 登录后 `data.token` |
    | `X-App-Version` | 建议 | `1.2.0` | 客户端版本号 |
    | `X-Platform` | 建议 | `ios` / `android` | 运行平台 |
    | `X-Device-Id` | 建议 | `A1B2C3D4-...` | 设备唯一标识 |
    | `X-Request-Id` | 建议 | UUID | 请求链路 ID |
    | `Accept-Language` | 可选 | `zh-CN` | 错误文案语言 |

    ---

    ## 三、统一返回格式

    HTTP 状态码表示传输层；**业务成败看 Body 中 `code`（string）**。

    ```json
    { "code": "0", "message": "ok", "data": {} }
    ```

    | 字段 | 类型 | 说明 |
    |------|------|------|
    | `code` | string | `"0"` 成功；非 `"0"` 业务失败 |
    | `message` | string | 提示文案 |
    | `data` | object / array / null | 业务数据 |

...

    ---

    ## 四、接口总览

    ### 4.1 应用配置
    | 方法 | 路径 | 说明 |
...

    ### 4.2 认证
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [认证]
      summary: 密码登录
      description: |
        | 项 | 值 |
        |----|-----|
        | 请求类型 | POST |
        | 是否鉴权 | **否**（勿传 Authorization） |
        | 请求头 | 仅需第二节中 POST 通用头（Content-Type 等），**不需要** Authorization |

        **Body 参数（JSON）**

        | 字段 | 类型 | 必填 | 说明 |
        |------|------|------|------|
        | phone | string | 是 | 手机号 |
        | password | string | 是 | 密码（至少 6 位） |

        **返回 data 字段**

        | 字段 | 类型 | 说明 |
        |------|------|------|
        | token | string | 访问令牌，后续 Header 使用 |
        | userId | string | 用户 Id |
        | username | string | 用户名 / 昵称 |
        | phone | string | 脱敏手机号，如 138****8000 |
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, password]
              properties:
                phone:
                  type: string
                  example: "13800138000"
                password:
                  type: string
                  example: "abc123456"
            example:
              phone: "13800138000"
              password: "abc123456"
      responses:
        '200':
          description: code 为 "0" 时登录成功
          content:
            application/json:
              schema:
                type: object
                required: [code, message]
                properties:
                  code:
                    type: string
                    example: "0"
                  message:
                    type: string
                    example: ok
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      userId:
                        type: string
                      username:
                        type: string
                      phone:
                        type: string
              example:
                code: "0"
                message: ok
                data:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  userId: "90001"
                  username: 王老板
                  phone: "138****8000"
  /auth/register:
    post:
      tags: [认证]
      summary: 注册
      description: |
        | 项 | 值 |
        |----|-----|
        | 请求类型 | POST |
        | 是否鉴权 | **否**（勿传 Authorization） |

        **Body 参数（JSON）**

        | 字段 | 类型 | 必填 | 说明 |
        |------|------|------|------|
        | phone | string | 是 | 手机号 |
        | password | string | 是 | 密码（至少 6 位） |
        | username | string | 否 | 昵称，未填时系统自动生成 |

        **返回 data 字段**：同 POST /auth/login
      operationId: authRegister
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, password]
              properties:
                phone:
                  type: string
                  example: "13900139000"
                password:
                  type: string
                  example: "abc123456"
                username:
                  type: string
                  example: "新老板"
            example:
              phone: "13900139000"
              password: "abc123456"
              username: "新老板"
      responses:
        '200':
          description: code 为 "0" 时注册成功并自动登录
          content:
            application/json:
              example:
                code: "0"
                message: ok
                data:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  userId: "90002"
                  username: 新老板
                  phone: "139****9000"
...
  /stores/{storeId}/menu/items/{itemId}:
    get:
      tags: [菜单菜品]
      summary: 菜品详情
      description: |
        | 项 | 值 |
        |----|-----|
        | 请求类型 | GET |
        | 是否鉴权 | **是** |

        **路径参数**

        | 字段 | 类型 | 说明 |
        |------|------|------|
        | storeId | integer | 门店 Id |
        | itemId | integer | 菜品 Id |

        **返回 data 字段**

        | 字段 | 类型 | 说明 |
        |------|------|------|
        | id | integer | 菜品 Id |
        | categoryId | integer | 分类 Id |
        | categoryName | string | 分类名称 |
        | name | string | 菜品名 |
        | description | string | 描述 |
        | price | integer | 售价（分） |
        | originalPrice | integer | 划线价（分） |
        | unit | string | 单位，如「份」 |
        | imageUrl | string | 图片 URL |
        | tags | string[] | 标签 |
        | spicyLevel | integer | 0不辣 1微 2中 3重 |
        | stock | integer | -1 不限库存 |
        | status | string | on_sale / off_sale |
        | sort | integer | 排序 |
        | createdAt | string | 创建时间 |
        | updatedAt | string | 更新时间 |
      operationId: getMenuItemDetail
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              example:
                code: "0"
                message: ok
                data:
                  id: 50001
                  categoryId: 1001
                  categoryName: 招牌热菜
                  name: 秘制红烧牛肉面
                  price: 3800
                  status: off_sale
                  sort: 100
                  createdAt: "2026-05-29 10:00:00"
  /stores/{storeId}/menu/items/save:
    post:
      tags: [菜单菜品]
      summary: 新增或编辑菜品
      description: |
        | 项 | 值 |
        |----|-----|
        | 请求类型 | POST |
        | 是否鉴权 | **是** |

        **路径参数**：storeId

        **Body 参数**

        | 字段 | 类型 | 必填 | 说明 |
        |------|------|------|------|
        | id | integer | 否 | 有=编辑，无=新增（新增默认 off_sale） |
        | categoryId | integer | 新增必填 | 分类 Id |
        | name | string | 新增必填 | 菜品名 |
        | description | string | 否 | 描述 |
        | price | integer | 新增必填 | 售价（分） |
        | originalPrice | integer | 否 | 划线价（分） |
        | unit | string | 否 | 默认「份」 |
        | imageUrl | string | 否 | 图片 |
        | tags | string[] | 否 | 标签 |
        | spicyLevel | integer | 否 | 0~3 |
        | stock | integer | 否 | 默认 -1 |
        | sort | integer | 否 | 排序 |

        **返回 data**：完整菜品对象（同详情）
      operationId: saveMenuItem
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            example:
              categoryId: 1001
              name: 秘制红烧牛肉面
              price: 3800
              description: 每日现熬牛骨汤
              spicyLevel: 1
              stock: -1
      responses:
        '200':
          content:
            application/json:
              example:
                code: "0"
                message: ok
                data:
                  id: 50001
                  name: 秘制红烧牛肉面
                  price: 3800
                  status: off_sale
  /stores/{storeId}/menu/items/remove:
    post:
      tags: [菜单菜品]
      summary: 删除菜品
      description: |
        | 项 | 值 |
        |----|-----|
        | 请求类型 | POST |
        | 是否鉴权 | **是** |

        **路径参数**：storeId

        **Body 参数**

        | 字段 | 类型 | 必填 | 说明 |
        |------|------|------|------|
        | itemId | integer | 是 | 菜品 Id |

        **返回 data**：成功为 null；上架中删除 code=4002
      operationId: removeMenuItem
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            example:
              itemId: 50001
      responses:
        '200':
          content:
            application/json:
              example:
                code: "0"
                message: ok
                data: null
  /stores/{storeId}/menu/items/status:
    post:
      tags: [菜单菜品]
      summary: 上架 / 下架
      description: |
        | 项 | 值 |
        |----|-----|
        | 请求类型 | POST |
        | 是否鉴权 | **是** |

        **路径参数**：storeId

        **Body 参数**

        | 字段 | 类型 | 必填 | 说明 |
        |------|------|------|------|
        | itemId | integer | 是 | 菜品 Id |
        | status | string | 是 | on_sale=上架 off_sale=下架 |

        **返回 data**：更新后的菜品对象
      operationId: updateMenuItemStatus
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            example:
              itemId: 50001
              status: on_sale
      responses:
        '200':
          content:
            application/json:
              example:
                code: "0"
                message: ok
                data:
                  id: 50001
                  status: on_sale
  /stores/{storeId}/menu/items/sort:
    post:
      tags: [菜单菜品]
      summary: 批量调整排序
      description: |
        | 项 | 值 |
        |----|-----|
        | 请求类型 | POST |
        | 是否鉴权 | **是** |

        **路径参数**：storeId

        **Body 参数**

        | 字段 | 类型 | 必填 | 说明 |
        |------|------|------|------|
        | itemIds | integer[] | 是 | 按展示顺序排列的菜品 Id 列表 |

        **返回 data**：成功为 null
      operationId: batchSortMenuItems
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            example:
              itemIds: [50003, 50001, 50002]
      responses:
        '200':
          content:
            application/json:
              example:
                code: "0"
                message: ok
                data: null
...
  /stores/{storeId}/menu/combos/save:
    post:
      tags: [菜单套餐]
      summary: 保存套餐（名称 + 价格 + 菜品）
      description: |
        | 项 | 值 |
        |----|-----|
        | 请求类型 | POST |
        | 是否鉴权 | **是** |

        无 `id` 为新增，有 `id` 为编辑。可同时保存套餐名称、售价及包含菜品。

        **Body 参数**

        | 字段 | 类型 | 必填 | 说明 |
        |------|------|------|------|
        | id | integer | 否 | 有=编辑，无=新增 |
        | name | string | 是 | 套餐名称（同店唯一） |
        | price | integer | 否 | 套餐售价（分），默认 0 |
        | itemIds | integer[] | 否 | 包含的菜品 Id 列表 |

        **返回 data**：同套餐详情
      operationId: saveMenuCombo
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: 商务午餐
              price: 5800
              itemIds: [50001, 50002]
      responses:
        '200':
          content:
            application/json:
              example:
                code: "0"
                message: ok
                data:
                  id: 2001
                  name: 商务午餐
                  price: 5800
                  items:
                    - id: 50001
                      categoryId: 1001
                      categoryName: 招牌热菜
                      name: 秘制红烧牛肉面
                      price: 3800
                      status: on_sale
  /stores/{storeId}/menu/combos/{comboId}:
    get:
      tags: [菜单套餐]
      summary: 套餐详情（价格 + 包含菜品）
      description: |
        | 项 | 值 |
        |----|-----|
        | 请求类型 | GET |
        | 是否鉴权 | **是** |

        根据套餐 Id 获取具体内容，含套餐售价及所含菜品列表。

        **路径参数**

        | 字段 | 类型 | 说明 |
        |------|------|------|
        | storeId | integer | 门店 Id |
        | comboId | integer | 套餐 Id |

        **返回 data 字段**

        | 字段 | 类型 | 说明 |
        |------|------|------|
        | id | integer | 套餐 Id |
        | name | string | 套餐名称 |
        | price | integer | 套餐售价（分） |
        | items | array | 包含的菜品列表 |

        **items[] 元素**

        | 字段 | 类型 | 说明 |
        |------|------|------|
        | id | integer | 菜品 Id |
        | categoryId | integer | 分类 Id |
        | categoryName | string | 分类名称 |
        | name | string | 菜品名称 |
        | price | integer | 菜品单价（分） |
        | status | string | on_sale / off_sale |
      operationId: getMenuComboDetail
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: comboId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              example:
                code: "0"
                message: ok
                data:
                  id: 2001
                  name: 商务午餐
                  price: 5800
                  items:
                    - id: 50001
                      categoryId: 1001
                      categoryName: 招牌热菜
                      name: 秘制红烧牛肉面
                      price: 3800
                      status: on_sale
                    - id: 50002
                      categoryId: 1002
                      categoryName: 饮品
                      name: 冰柠檬茶
                      price: 1200
                      status: on_sale
  /stores/{storeId}/menu/combos/{comboId}/save-content:
    post:
      tags: [菜单套餐]
      summary: 保存套餐内容（价格 + 菜品）
      description: |
        | 项 | 值 |
        |----|-----|
        | 请求类型 | POST |
        | 是否鉴权 | **是** |

        **Body 参数**

        | 字段 | 类型 | 必填 | 说明 |
        |------|------|------|------|
        | price | integer | 是 | 套餐售价（分） |
        | itemIds | integer[] | 是 | 包含的菜品 Id 列表（至少 1 个） |

        **返回 data**：同套餐详情
      operationId: saveMenuComboContent
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: comboId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            example:
              price: 5800
              itemIds: [50001, 50002]
      responses:
        '200':
          content:
            application/json:
              example:
                code: "0"
                message: ok
                data:
                  id: 2001
                  name: 商务午餐
                  price: 5800
                  items:
                    - id: 50001
                      categoryId: 1001
                      categoryName: 招牌热菜
                      name: 秘制红烧牛肉面
                      price: 3800
                      status: on_sale
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 见第二节 Authorization；Token 来自 POST /auth/login 的 data.token